backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
supabase==2.28.2
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
httpx==0.27.0
google-generativeai==0.8.0
